move createComponentSSRTest.ts
