+++ src/components/card/Card.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import Card from "./Card";
 import { render } from "../../test-utils/render";
-import { createComponentSSRTest } from "../../createComponentSSRTest";
+import { createComponentSSRTest } from "../../test-utils/createComponentSSRTest";
 
 describe("Card", () => {
   it("renders without crashing", () => {
