+++ src/components/codefield/CodeField.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import CodeField from "./CodeField";
 import { render } from "../../test-utils/render";
-import { createComponentSSRTest } from "../../createComponentSSRTest";
+import { createComponentSSRTest } from "../../test-utils/createComponentSSRTest";
 
 describe("CodeField", () => {
   it("renders without crashing", () => {
