+++ src/components/navigation-bar/NavigationBar.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import NavigationBar from "./NavigationBar";
 import { render } from "../../test-utils/render";
-import { createComponentSSRTest } from "../../createComponentSSRTest";
+import { createComponentSSRTest } from "../../test-utils/createComponentSSRTest";
 
 export const items = [
   {
