+++ src/components/radio/Radio.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import Radio from "./Radio";
 import { render } from "../../test-utils/render";
-import { createComponentSSRTest } from "../../createComponentSSRTest";
+import { createComponentSSRTest } from "../../test-utils/createComponentSSRTest";
 
 describe("Radio", () => {
   it("renders without crashing", () => {
