+++ src/components/spinner/Spinner.test.tsx
@@ -2,7 +2,7 @@ import { screen } from "@testing-library/react";
 import Spinner from "./Spinner";
 import { SPINNER_SIZE } from "./types";
 import { render } from "../../test-utils/render";
-import { createComponentSSRTest } from "../../createComponentSSRTest";
+import { createComponentSSRTest } from "../../test-utils/createComponentSSRTest";
 
 describe("Spinner", () => {
   it("renders without crashing", () => {
